+++ home-manager/modules/zsh.nix
@@ -1,7 +1,7 @@
 ####### Configuration for Zsh ##################################################
 ##                                                                            ##
 ## * Configure the prompt to be blue in Nix shells                            ##
-## * Enable Oh My Zsh with nix-shell and sudo plugins                         ##
+## * Optionally enable Oh My Zsh with nix-shell and sudo plugins              ##
 ## * Use the Bazik Oh My Zsh theme from confkit                               ##
 ## * Provide the confkit.zsh.plugins configuration option to select which     ##
 ##   plugins from confkit/zsh are to be installed $HOME/.zsh                  ##
@@ -12,7 +12,7 @@
 
 let
   inherit (builtins) readFile;
-  inherit (lib) mkOption mkEnableOption mkIf mkMerge types;
+  inherit (lib) mkOption mkEnableOption mkIf mkDefault mkMerge types;
   cfg = config.confkit.zsh;
 in
 
@@ -31,6 +31,13 @@ in
         defined in confkit/zsh/nix.zsh, set confkit.zsh.plugins = [ "nix" ];
       '';
     };
+
+    ohMyZsh = mkOption {
+      type = types.bool;
+      default = false;
+      example = true;
+      description = "Wether to enable Oh My Zsh.";
+    };
   };
 
   config = mkIf cfg.enable (mkMerge [
@@ -39,16 +46,18 @@ in
         enable = true;
         initExtra = readFile ../../zsh/config/home_init.zsh;
 
-        oh-my-zsh = {
+        oh-my-zsh = mkIf cfg.ohMyZsh {
           enable = true;
-          custom = "$HOME/.zsh-custom";
-          theme = "bazik";
+          custom = mkDefault "$HOME/.zsh-custom";
+          theme = mkDefault "bazik";
           plugins = [ "nix-shell" "sudo" ];
         };
       };
 
-      home.file.".zsh-custom/themes/bazik.zsh-theme".source =
-        ../../zsh/themes/bazik.zsh-theme;
+      home.file = mkIf cfg.ohMyZsh {
+        ".zsh-custom/themes/bazik.zsh-theme".source =
+          ../../zsh/themes/bazik.zsh-theme;
+      };
     }
 
     {